Avoids recalculating murmur hashes for keys for each bloom filter pass by caching hashes for decorated keys supplied by the partitioners. To avoid increasing the memory usage, cached hashes are not saved in decorated keys stored in memtables.
As the hash is calculated during decoration of a key, the patch might
cause slowdown for some partitioners when no bloom filters are in use
(i.e. MemTable-only scenarios). This is not the case for the Murmur3
partitioner, where a single pass is now used for calculating both the
bloom filter hash and token.