Variable shard count, using base shard count and target sstable size
Changes sharding to select a number of shards to split an sstable based
on its expected size. A minimum shard count is used when the size of a
level is too small, and for any higher level a power-of-two multiple of
the base count is selected that splits the expected level size into
shards whose size is targeted to be between sqrt(0.5) * target and
sqrt(2) * target.

To achieve this, ShardManager is changed to generate shard split points
on the fly for any given shard count. SSTables are always split on the
selected shard boundaries to maximize efficiency.

The sharding is completely independent of compaction selection including
the choice of level fanout factors, thus just as suitable for any
compaction configuration (i.e. cannot cause inefficiencies on
reconfiguration). Because the next sharding level only splits the
previous one in the middle, even if we end up mixing shard counts within
a compaction level, the bucket processing will easily identify the split
points it needs to work with (as long as the overlap inclusion method
is at least SINGLE) and will in-effect treat non-overlapping subshards
as a single run.